Reorganize search route
