+++ app/routes/api/api.router.js
@@ -6,7 +6,7 @@ const albumRouter = require("./album/router");
 const artistRouter = require("./artist.api.router");
 const crateRouter = require("./crate/router");
 const playlistRouter = require("./playlist/router");
-const searchRouter = require("./search.api.router");
+const searchRouter = require("./search/router");
 const tokenRouter = require("./token.api.router");
 const { sendErrorCode } = require("./errors");
 const authorizeWithToken = passport.authenticate("jwt", { session: false });
@@ -29,15 +29,10 @@ router.get(
 
 // routes
 router.use("/album", authorizeWithToken, albumRouter);
-
 router.use("/artist", authorizeWithToken, artistRouter);
-
 router.use("/crate", authorizeWithToken, crateRouter);
-
 router.use("/playlist", authorizeWithToken, playlistRouter);
-
 router.use("/search", authorizeWithToken, searchRouter);
-
 router.use("/token", tokenRouter);
 
 // error handling
